+++ Demo/src/main/java/com/braintreepayments/demo/PendingRequestStore.java
@@ -4,7 +4,7 @@ import android.content.Context;
 import android.content.SharedPreferences;
 
 import com.braintreepayments.api.localpayment.LocalPaymentPendingRequest;
-import com.braintreepayments.api.PayPalPendingRequest;
+import com.braintreepayments.api.paypal.PayPalPendingRequest;
 import com.braintreepayments.api.sepadirectdebit.SEPADirectDebitPendingRequest;
 import com.braintreepayments.api.venmo.VenmoPendingRequest;