+++ GooglePay/src/main/java/com/braintreepayments/api/googlepay/GooglePayCardNonce.java
@@ -9,7 +9,7 @@ import androidx.annotation.NonNull;
 import com.braintreepayments.api.core.PaymentMethodNonce;
 import com.braintreepayments.api.core.PostalAddress;
 import com.braintreepayments.api.sharedutils.Json;
-import com.braintreepayments.api.PayPalAccountNonce;
+import com.braintreepayments.api.paypal.PayPalAccountNonce;
 import com.braintreepayments.api.card.BinData;
 
 import org.json.JSONException;