+++ PayPal/src/main/java/com/braintreepayments/api/PayPalInternalClientCallback.java
@@ -1,7 +1,0 @@
-package com.braintreepayments.api;
-
-import androidx.annotation.Nullable;
-
-interface PayPalInternalClientCallback {
-    void onResult(@Nullable PayPalPaymentAuthRequestParams paymentAuthRequest, @Nullable Exception error);
-}