+++ PayPal/src/main/java/com/braintreepayments/api/PayPalInternalTokenizeCallback.java
@@ -1,8 +1,0 @@
-package com.braintreepayments.api;
-
-import androidx.annotation.Nullable;
-
-interface PayPalInternalTokenizeCallback {
-
-    void onResult(@Nullable PayPalAccountNonce payPalAccountNonce, @Nullable Exception error);
-}