+++ PayPal/src/main/java/com/braintreepayments/api/paypal/PayPalInternalTokenizeCallback.java
@@ -1,0 +1,8 @@
+package com.braintreepayments.api.paypal;
+
+import androidx.annotation.Nullable;
+
+interface PayPalInternalTokenizeCallback {
+
+    void onResult(@Nullable PayPalAccountNonce payPalAccountNonce, @Nullable Exception error);
+}