+++ PayPal/src/test/java/com/braintreepayments/api/paypal/PayPalPaymentResourceUnitTest.java
@@ -1,0 +1,34 @@
+package com.braintreepayments.api.paypal;
+
+import org.json.JSONException;
+import org.json.JSONObject;
+import org.junit.Test;
+
+import static org.junit.Assert.*;
+
+import com.braintreepayments.api.paypal.PayPalPaymentResource;
+
+public class PayPalPaymentResourceUnitTest {
+
+    @Test
+    public void fromJson_parsesRedirectUrlFromOneTimePaymentResource() throws JSONException {
+        String oneTimePaymentJson = new JSONObject()
+                .put("paymentResource", new JSONObject()
+                        .put("redirectUrl", "www.example.com/redirect")
+                ).toString();
+
+        PayPalPaymentResource sut = PayPalPaymentResource.fromJson(oneTimePaymentJson);
+        assertEquals("www.example.com/redirect", sut.getRedirectUrl());
+    }
+
+    @Test
+    public void fromJson_parsesRedirectUrlFromBillingAgreementPaymentResource() throws JSONException {
+        String billingAgreementJson = new JSONObject()
+                .put("agreementSetup", new JSONObject()
+                        .put("approvalUrl", "www.example.com/redirect")
+                ).toString();
+
+        PayPalPaymentResource sut = PayPalPaymentResource.fromJson(billingAgreementJson);
+        assertEquals("www.example.com/redirect", sut.getRedirectUrl());
+    }
+}